Native

+++ rayTracer/src/emscriptenRunner.h
@@ -1,11 +1,11 @@
 #ifdef __EMSCRIPTEN__
 
-#include <vector>
-#include <memory>
-#include <utility>
-#include <string>
 #include <algorithm>
+#include <memory>
 #include <random>
+#include <string>
+#include <utility>
+#include <vector>
 
 #include <glm/glm.hpp>
 
@@ -13,39 +13,38 @@
 #include <emscripten/val.h>
 
 #include "camera.h"
+#include "objectLoader.h"
 #include "renderer.h"
 #include "world.h"
-#include "objectLoader.h"
 
-class EmscriptenRunner
-{
+class EmscriptenRunner {
 private:
-    static const int PIXELS_PER_BATCH = 20000;
-    World world;
-    std::shared_ptr<Camera> camera;
-    Renderer renderer;
-    std::vector<std::pair<int, int>> pixelsToRender;
+  static const int PIXELS_PER_BATCH = 20000;
+  World world;
+  std::shared_ptr<Camera> camera;
+  Renderer renderer;
+  std::vector<std::pair<int, int>> pixelsToRender;
 
-    int sqrtRaysPerPixel;
-    float halfSubPixelSize;
-    const float STEP_SIZE = 0.05f;
+  int sqrtRaysPerPixel;
+  double halfSubPixelSize;
+  const double STEP_SIZE = 0.05f;
 
-    std::random_device rd;
-    std::mt19937 g;
+  std::random_device rd;
+  std::mt19937 g;
 
-    void moveCamera(float posChange);
+  void moveCamera(double posChange);
 
 public:
-    EmscriptenRunner(/* args */);
-    void init(const std::string &sceneDesc);
-    ~EmscriptenRunner();
+  EmscriptenRunner(/* args */);
+  void init(const std::string &sceneDesc);
+  ~EmscriptenRunner();
 
-    emscripten::val renderToRGBA();
-    int getHeight();
-    int getWidth();
-    void moveLeft();
-    void moveRight();
-    bool done();
+  emscripten::val renderToRGBA();
+  int getHeight();
+  int getWidth();
+  void moveLeft();
+  void moveRight();
+  bool done();
 };
 
 #endif
