+++ rayTracer/src/pattern.cpp
@@ -1,8 +1,8 @@
 #include "pattern.h"
 
 Pattern::Pattern() {
-  this->transform = glm::mat4(1.f);
-  this->inverseTransform = glm::mat4(1.f);
+  this->transform = glm::dmat4(1.0);
+  this->inverseTransform = glm::dmat4(1.0);
 }
 
 Pattern::Pattern(const Pattern &pattern) {
@@ -10,12 +10,12 @@ Pattern::Pattern(const Pattern &pattern) {
   this->inverseTransform = pattern.inverseTransform;
 }
 
-void Pattern::setTransform(glm::mat4 transform) {
+void Pattern::setTransform(glm::dmat4 transform) {
   this->transform = transform;
   // this->inverseTransform = glm::affineInverse(this->transform);
 }
 
-void Pattern::multiplyTransform(glm::mat4 &transform) {
+void Pattern::multiplyTransform(glm::dmat4 &transform) {
   this->transform = transform * this->transform;
   // this->inverseTransform = glm::affineInverse(this->transform);
 }
@@ -28,7 +28,7 @@ Pattern::~Pattern() {}
 
 ColourPattern::ColourPattern() {}
 
-ColourPattern::ColourPattern(glm::vec3 colourA, glm::vec3 colourB) : Pattern() {
+ColourPattern::ColourPattern(glm::dvec3 colourA, glm::dvec3 colourB) : Pattern() {
   this->colourA = colourA;
   this->colourB = colourB;
 }
@@ -38,7 +38,7 @@ ColourPattern::ColourPattern(const ColourPattern &colouredPattern) : Pattern() {
   this->colourB = colouredPattern.colourB;
 }
 
-void ColourPattern::setColour(glm::vec3 colour, int index) {
+void ColourPattern::setColour(glm::dvec3 colour, int index) {
   if (index == 0)
     this->colourA = colour;
   else if (index == 1)
